Simplify the code examples for UIDispatcher.Post
The examples of how to use UIDispatcher.Post method were simplified so
now there's no separate method for handling the file logic, but it is
embedded into the lambda body instead.
This is to avoid the "godd practice" inclusions, so there's no explicit
requirements of how the code should be written.