Describe `UIDispatcher`'s usage in the Using Asynchronous Windows APIs document (#428)
* Describe the need and reason for using UIDispatcher

* Describe UIDispatcher usage in C# code

* Describe UIDispatcher usage in C++ code

* Cover the review comments

* Cover review remarks

* Simplify the code examples for UIDispatcher.Post

The examples of how to use UIDispatcher.Post method were simplified so
now there's no separate method for handling the file logic, but it is
embedded into the lambda body instead.
This is to avoid the "godd practice" inclusions, so there's no explicit
requirements of how the code should be written.

* Correct the code formatting in examples

* Fix: Section not found in native-modules-async

The typo corrected in this commit fixes the CI check for valid links.

* Fix: typo in ReactInitializer link

* Update docs/native-modules-async.md

Co-authored-by: Alexander Sklar <asklar@microsoft.com>

Co-authored-by: Alexander Sklar <asklar@microsoft.com>