Separate `create-react-native-module` as another option
Official React Native instructions use react-native-builder-bob from
Callstack.io company to create the native module project, while the
example presented below it uses other community library.
It is essential to emphasize that these are not the same ways to create
the project.
This commit adds "or" to ensure that the reader will be aware of those
options.