Update AppTheme Docs for 0.65 (#487)
* Update AppTheme Docs for 0.65

Everything in AppTheme that overlapped with Appearance was removed. Remove the APIs, and rewrite the sample to not use them. We convert from a class component to function component when updating the sample. The methods are also exposing as property getters instead of methods, so we change the examples there.

* correct initial state

* missing import

* actually run it through ESLint

* update