Re-cut 0.65 Docs
versioned_docs for 0.65 seems to be missing some recent commits that were intended to go into 0.65 docs. This commit re-cuts the earlier docs, then reconciles with the changes that happened when cutting.

The 0.65 directory seems to also be missing a couple of whole files, e.g. apptheme-api-windows, which is present in main and older docs. The current site seems to be falling back to 0.64 docs in these cases.