Add specs for sample native modules (#573)
Fix a couple of issues with the current native module samples.

On the JS side they were exporting undefined. Now they are exporting an object containing the two NativeModules that they implement. The native modules are also now exporting JS type information. This is flow for now, since this is what the packages are using currently. This will also be able to be used with the codegen package in the future to generate turbomodules.

Co-authored-by: Jon Thysell (JAUNTY) <jthysell@microsoft.com>