Fix module builder bugs in "Native Modules (Advanced)" page
The sample C# code given in the Native Modules (Advanced) page is
incorrect. First, there is no `IReactModuleBuilder.SetName` method
(it might be left over from an earlier iteration of the API).

Also, when passing an `IJSValueWriter` to a `MethodResultCallback`,
the writer must contain an array of arguments to pass to the callback.
Even if the callback has only one argument, it must be wrapped in an
array for the (de)serialization to work.

Closes https://github.com/microsoft/react-native-windows/issues/10591