Refactor Workflows
This PR refactors the multiple separate CI and PR workflows in this repo into a single CI and PR workflow using reusable template workflows rather than copy/pasting shared jobs/steps. This should simplify maintenance of the samples.

The CI workflow is designed to build the full matrix of platform and configurations for each supported sample in the repo. It still runs the same time every week but can also be started from the GitHub Actions UI.

The PR workflow is designed to build a single flavor for each supported sample that was actually modified in the PR. If no samples (or the website) were modified, then the PR will simply pass. As such this refactor also ensures that all PRs comply with the repo's required PR checks.

This PR does not refactor the rssreader sample's workflows (as they don't build as-is).