Fix broken init command in documentation (#1013)
## Description

### Why
This PR fixes a insonsitency in the documentation and updates the init
command to the latest date because the current command is resulting in
the following error
```
npx react-native@latest init ReactNativeGithub --version 0.76.0

🚨️ The `init` command is deprecated.

- Switch to npx @react-native-community/cli init for the identical behavior.
- Refer to the documentation for information about alternative tools: https://reactnative.dev/docs/getting-started
Exiting...
```
Furthermore, I've removed the path level from the version value so the
latest patch will be used (0.76.7)
What is the motivation for this change? Add a few sentences describing
the context and overall goals of the pull request's commits.
###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/react-native-windows-samples/pull/1013)

---------

Co-authored-by: Saad Najmi <sanajmi@microsoft.com>