Add note about Old and New Architectures to relevant docs (#1029)
## Description

This PR adds a notice to the top of every relevant page that is either
clearly:

1. *Old Architecture Only*: doc is for Old Architecture features not
planned/supported in the New Arch
2. *Architecture Review Needed*: doc may or may not apply to in the new
architecture and therefore needs further review.

I've also added the appropriate tag on every page:


![Architecture](https://img.shields.io/badge/architecture-new_&_old-green)

![Architecture](https://img.shields.io/badge/architecture-new_only-blue)


![Architecture](https://img.shields.io/badge/architecture-old_only-yellow)


![Architecture](https://img.shields.io/badge/architecture-needs_review-red)

### Why
Lots of things change with the migration to the New Architecture, and we
want to give some warning about docs that may be out of date while we
sort out which to update, which to deprecate, etc.

Resolves #1014 

## Screenshots
N/A

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/react-native-windows-samples/pull/1029)