New Native Platform docs describing the `cpp-lib` template (#1032)
## Description

This PR creates new "Native Platform" docs to capture using the new `cpp-lib` template to add Windows support libraries for both Native Modules and Native Components. This PR also moves the existing "Native Module" docs to a "Legacy Docs" section on the sidebar - each is also given a link to direct devs toward the new docs.

### Why

The current "Native Module" docs may or may not be relevant with the new template - this PR is to move the focus off of those docs (without deleting them just yet) and have clean updated docs that describe how to use the new template.

Resolves #982
Resolves #999 

## Screenshots
N/A

### Why

The current "Native Module" docs may or may not be relevant with the new
template - this PR is to move the focus off of those docs (without
deleting them just yet) and have clean updated docs that describe how to
use the new template.

Resolves #982
Resolves #999 

## Screenshots
N/A

---------

Co-authored-by: Tatiana Kapos <tatianakapos@microsoft.com>