Update documentation to reflect New Architecture default (#1051)
## Description

This PR updates the documentation to reflect the New Architecture
default for new applications in 0.80, and also refreshes the
documentation on the differences between the architectures for RNW.

### Why
We are switching the default for new projects in 0.80. We are also
directing more folks to the "New vs. Old Architecture" page both within
the docs and from outside.

Resolves #1046

## Screenshots
N/A

###### Microsoft Reviewers: [Open in
CodeFlow](https://microsoft.github.io/open-pr/?codeflow=https://github.com/microsoft/react-native-windows-samples/pull/1051)